+++ .eslintrc
@@ -1,5 +1,8 @@
 root: true
 
+rules:
+  no-console: 1
+
 parserOptions:
   ecmaVersion: 2018
   sourceType: "module"
