--- js/schema.json
@@ -1,16 +1,0 @@
-[
-  {
-    "id": 1,
-    "url": "photos/1.jpeg",
-    "description": "Some photo",
-    "likes": 24,
-    "comments": [
-      {
-        "id": 135,
-        "avatar": "img/avatar-6.svg",
-        "message": "В целом всё неплохо. Но не всё.",
-        "name": "Артём"
-      }
-    ]
-  }
-]
